Extract Cachex code so that new caching strategies can be implemented
Note, I've called the class "CachexCache" which is horrendous and likely
temporary. The reasoning here is that we've not got a complex
folder/namespace structure yet so I'm just being. The intention is to
add an MnesiaCache as part of this PR and allow us to add others in the
future if we want. Ultimately, I'll most likely move this into a new
namespace Closex.Cache instead. I may even do that next...