Respect rate limits and retry once on find actions (#26)
We're hitting rate limits on (specifically) `find` operations,
triggering a sentry here: https://sentry.io/nested/elixir/issues/563686900/

This inspects and respsects the rate limit information we get back from
Close.io. If we hit a rate limit we will:

1. Inspect the `rate_reset` window (time until the window resets)
2. Wait for 1 second more than the window specifies
3. Retry once

We only retry once because if we're spamming their API, we probably
don't want to spam indefinitely if we continually hit rate limits. We
also want to get an alert telling us this is the case